+++ packages/template-retail-react-app/app/page-designer/assets/image-tile/index.jsx
