+++ packages/template-retail-react-app/app/page-designer/assets/image-with-text/index.jsx
