+++ packages/template-retail-react-app/app/page-designer/core/index.js
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2023, Salesforce, Inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: BSD-3-Clause
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+
+export * from './component'
+export * from './page'
+export * from './region'
+export * from './types'
