+++ packages/template-retail-react-app/app/page-designer/layouts/mobileGrid1r1c/index.jsx
@@ -7,12 +7,10 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import {SimpleGrid} from '@chakra-ui/react'
-import {Region} from '../../region'
-import {regionType} from '../../types'
+import {Region, regionType} from '../../core'
 
 /**
- * This layout component displays its children in a 2 row x 1 column grid on mobile
- * and a 1 row x 2 column grid on desktop.
+ * This layout component displays its children in a 1 x 1 grid on both mobile and desktop.
  *
  * @param {componentProps} props
  * @param {regionType []} props.regions - The page designer regions for this component.
@@ -20,19 +18,19 @@ import {regionType} from '../../types'
  * @param {string} props.typeId - A mapping of typeId's to react components representing the type.
  * @returns {React.ReactElement} - Grid component.
  */
-export const MobileGrid2r1c = ({regions}) => (
-    <SimpleGrid className="mobile-2r-1c" columns={{base: 1, sm: 2}}>
+export const MobileGrid1r1c = ({regions}) => (
+    <SimpleGrid className="mobile-1r-1c" columns={1}>
         {regions.map((region) => (
             <Region key={region.id} region={region} />
         ))}
     </SimpleGrid>
 )
 
-MobileGrid2r1c.displayName = 'MobileGrid2r1c'
+MobileGrid1r1c.displayName = 'MobileGrid1r1c'
 
-MobileGrid2r1c.propTypes = {
+MobileGrid1r1c.propTypes = {
     // Internally Provided
     regions: PropTypes.arrayOf(regionType).isRequired
 }
 
-export default MobileGrid2r1c
+export default MobileGrid1r1c
