+++ packages/template-retail-react-app/app/page-designer/layouts/mobileGrid2r3c/index.test.js
@@ -5,7 +5,7 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import React from 'react'
-import {renderWithProviders, withPageProvider} from '../../../../utils/test-utils'
+import {renderWithProviders, withPageProvider} from '../../../utils/test-utils'
 import MobileGrid2r3c from './index'
 
 test('MobileGrid2r3c renders without errors', () => {
