+++ packages/template-retail-react-app/app/page-designer/layouts/mobileGrid3r2c/index.jsx
@@ -7,12 +7,11 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import {SimpleGrid} from '@chakra-ui/react'
-import {Region} from '../../region'
-import {regionType} from '../../types'
+import {Region, regionType} from '../../core'
 
 /**
- * This layout component displays its children in a 2 row x 2 column grid on mobile
- * and 1 row x 4 column grid on desktop.
+ * This layout component displays its children in a 3 row x 2 column grid on mobile
+ * and 2 row x 3 column grid on desktop.
  *
  * @param {componentProps} props
  * @param {regionType []} props.regions - The page designer regions for this component.
@@ -20,19 +19,19 @@ import {regionType} from '../../types'
  * @param {string} props.typeId - A mapping of typeId's to react components representing the type.
  * @returns {React.ReactElement} - Grid component.
  */
-export const MobileGrid2r2c = ({regions}) => (
-    <SimpleGrid className="mobile-2r-2c" columns={{base: 2, sm: 4}}>
+export const MobileGrid3r2c = ({regions}) => (
+    <SimpleGrid className="mobile-3r-2c" columns={{base: 2, sm: 3}}>
         {regions.map((region) => (
             <Region key={region.id} region={region} />
         ))}
     </SimpleGrid>
 )
 
-MobileGrid2r2c.displayName = 'MobileGrid2r2c'
+MobileGrid3r2c.displayName = 'MobileGrid3r2c'
 
-MobileGrid2r2c.propTypes = {
+MobileGrid3r2c.propTypes = {
     // Internally Provided
     regions: PropTypes.arrayOf(regionType).isRequired
 }
 
-export default MobileGrid2r2c
+export default MobileGrid3r2c
