+++ packages/template-retail-react-app/app/page-designer/layouts/mobileGrid3r2c/index.test.js
@@ -5,7 +5,7 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import React from 'react'
-import {renderWithProviders, withPageProvider} from '../../../../utils/test-utils'
+import {renderWithProviders, withPageProvider} from '../../../utils/test-utils'
 import MobileGrid3r2c from './index'
 
 test('MobileGrid3r2c renders without errors', () => {
