+++ packages/template-retail-react-app/app/utils/test-utils.js
@@ -9,8 +9,7 @@ import {render} from '@testing-library/react'
 import {BrowserRouter as Router} from 'react-router-dom'
 import {ChakraProvider} from '@chakra-ui/react'
 import PropTypes from 'prop-types'
-import {PageContext} from '../components/experience/page'
-import {Region} from '../components/experience/region'
+import {PageContext, Region} from '../page-designer/core'
 
 import theme from '../theme'
 import CommerceAPI from '../commerce-api'
